Add cargo build scripts to compile Mesa.
The configure flags used disable everything possible
except for OSMesa (the offscreen software rendering
portion of Mesa).

If LLVM libraries are detected on the system, the llvmpipe
driver will be built.

However, for Servo testing, we prefer the softpipe driver anyway
since this is faster for CI testing (shader compile times
dominate otherwise when running 10,000 simple tests).

The softpipe driver is always built, regardless of whether
the llvmpipe driver is selected to build.